refactor(native-ffi): split FFI implementation modules
